Remove a duplicate file and add 3 lives to the hole game
